s/HttpRequestBuilder/HttpRequestFormatter/g
